+++ Changelog.md
@@ -2,6 +2,7 @@
 - Extra setting to configure stepInterval.
 - Optimized substeps calculation.
 - Optimized control loop, trigger pid faster after target changes.
+- Different PID settings for Boil/Mash.
 
 # Version 1.3.1
 - Bugfix: One cannot delete steps.